# backend/dist/swagger.json
{
	"openapi": "3.0.0",
	"components": {
...
		"schemas": {
...
			"TimelineEventResponseType": {
				"properties": {
					"_id": {
						"type": "string"
					},
					"projectId": {
						"type": "string"
					},
					"title": {
						"type": "string"
					},
					"order": {
						"type": "number",
						"format": "double"
					},
					"description_html": {
						"type": "string"
					},
					"imageUrl": {
						"type": "string"
					},
					"createdAt": {
						"type": "string",
						"format": "date-time"
					},
					"updatedAt": {
						"type": "string",
						"format": "date-time"
					}
				},
				"required": [
					"_id",
					"projectId",
					"title",
					"order",
					"description_html",
					"imageUrl",
					"createdAt",
					"updatedAt"
				],
				"type": "object",
				"additionalProperties": false
			},
			"CreateTimelineEventInput": {
				"properties": {
					"imageUrl": {
						"type": "string"
					},
					"description_html": {
						"type": "string"
					},
					"order": {
						"type": "number",
						"format": "double"
					},
					"title": {
						"type": "string"
					},
					"projectId": {
						"type": "string"
					}
				},
				"type": "object"
			},
			"UpdateTimelineEventInput": {
				"properties": {
					"imageUrl": {
						"type": "string"
					},
					"description_html": {
						"type": "string"
					},
					"order": {
						"type": "number",
						"format": "double"
					},
					"title": {
						"type": "string"
					}
				},
				"type": "object"
			},
...
		},
		"securitySchemes": {
			"jwt": {
				"type": "apiKey",
				"name": "Authorization",
				"in": "header",
				"description": "Token JWT no formato 'Bearer {token}'"
			}
		}
	},
	"info": {
		"title": "backend",
		"version": "1.0.0",
		"description": "Backend para o TCC Usina Guara",
		"license": {
			"name": "ISC"
		},
		"contact": {
			"name": "Laysa Bernardes"
		}
	},
	"paths": {
...
		"/projects/{projectId}/timeline": {
			"post": {
				"operationId": "CreateTimelineEvent",
				"responses": {
					"201": {
						"description": "Created",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/TimelineEventResponseType"
								}
							}
						}
					}
				},
				"tags": [
					"Timeline"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "projectId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/CreateTimelineEventInput"
							}
						}
					}
				}
			},
			"get": {
				"operationId": "GetTimelineEventsForProject",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/TimelineEventResponseType"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"tags": [
					"Timeline"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "projectId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/timeline/{timelineEventId}": {
			"get": {
				"operationId": "GetTimelineEventById",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/TimelineEventResponseType"
								}
							}
						}
					},
					"404": {
						"description": "Not Found"
					}
				},
				"tags": [
					"Timeline"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "timelineEventId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			},
			"put": {
				"operationId": "UpdateTimelineEvent",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/TimelineEventResponseType"
								}
							}
						}
					},
					"404": {
						"description": "Not Found"
					}
				},
				"tags": [
					"Timeline"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "timelineEventId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/UpdateTimelineEventInput"
							}
						}
					}
				}
			},
			"delete": {
				"operationId": "DeleteTimelineEvent",
				"responses": {
					"204": {
						"description": "No Content"
					},
					"404": {
						"description": "Not Found"
					}
				},
				"tags": [
					"Timeline"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "timelineEventId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
...
	},
	"servers": [
		{
			"url": "/"
		}
	]
}